gnu cc improvements
git-svn-id: http://svn.cip4.org/svnroot/trunk@6663 25ee8a6d-831b-0410-907a-ff8528d65d49

+++ src/jdf/lang/WStringBase.h
@@ -1005,7 +1005,7 @@ int WStringBase<T>::indexOf(const WStringBase<T>& str, int fromIndex) const
 			i++;
 		
 		if (i > max) 
-			return WStringBase::npos;
+			return (int)WStringBase::npos;
 		
 		/* Found first character, now look at the rest of v2 */
 		int j = i + 1;
